Divide each interface in separated file following C# solution
