+++ sdk/typescript/libraries/botbuilder-skills/src/models/manifest/action.ts
@@ -1,0 +1,11 @@
+/**
+ * Copyright(c) Microsoft Corporation.All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { IActionDefinition } from "./";
+
+export interface IAction {
+    id: string;
+    definition: IActionDefinition;
+}
