+++ sdk/typescript/libraries/botbuilder-skills/src/models/manifest/authenticationConnection.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright(c) Microsoft Corporation.All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+/**
+ * Describes an Authentication connection that a Skill requires for operation.
+ */
+export interface IAuthenticationConnection {
+    id: string;
+    serviceProviderId: string;
+    scopes: string;
+}
