+++ sdk/typescript/libraries/botbuilder-skills/src/models/manifest/event.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright(c) Microsoft Corporation.All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export interface IEvent {
+    name: string;
+}
