+++ sdk/typescript/libraries/botbuilder-skills/src/models/manifest/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Copyright(c) Microsoft Corporation.All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export * from './action';
+export * from './actionDefinition';
+export * from './authenticationConnection';
+export * from './event';
+export * from './skillManifest';
+export * from './slot';
+export * from './triggers';
+export * from './utterance';
+export * from './utteranceSource';
