+++ sdk/typescript/libraries/botbuilder-skills/src/models/manifest/slot.ts
@@ -1,0 +1,12 @@
+/**
+ * Copyright(c) Microsoft Corporation.All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+/**
+ * Slot definition for a given Action within a Skill manifest.
+ */
+export interface ISlot {
+    name: string;
+    types: string[];
+}
