+++ sdk/typescript/libraries/botbuilder-skills/src/models/manifest/triggers.ts
@@ -1,0 +1,15 @@
+/**
+ * Copyright(c) Microsoft Corporation.All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { IEvent, IUtterance, IUtteranceSources } from "./";
+
+/**
+ * Definition of the triggers for a given action within a Skill.
+ */
+export interface ITriggers {
+    utterances?: IUtterance[];
+    utteranceSources: IUtteranceSources[];
+    events?: IEvent[];
+}
