+++ sdk/typescript/libraries/botbuilder-skills/src/models/manifest/utteranceSource.ts
@@ -1,0 +1,12 @@
+/**
+ * Copyright(c) Microsoft Corporation.All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+/**
+ * Source of utterances for a given locale which form part of an Action within a manifest.
+ */
+export interface IUtteranceSources {
+    locale: string;
+    source: string[];
+}
